requirements.txt
alembic==1.2.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BasicAuth==0.2.0
flask-restplus==0.13.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.2.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
